Extract power sources graph card series generation into a pure module
