+++ src/panels/lovelace/cards/energy/hui-power-sources-graph-card.ts
@@ -1,4 +1,4 @@
-import { endOfToday, isSameDay, isToday, startOfToday } from "date-fns";
+import { endOfToday, isToday, startOfToday } from "date-fns";
 import type { HassConfig, UnsubscribeFunc } from "home-assistant-js-websocket";
 import type { PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
@@ -6,27 +6,23 @@ import { customElement, property, state } from "lit/decorators";
 import { classMap } from "lit/directives/class-map";
 import memoizeOne from "memoize-one";
 import type { LineSeriesOption } from "echarts/charts";
-import { LinearGradient } from "../../../../resources/echarts/echarts";
 import "../../../../components/chart/ha-chart-base";
-import { computeYAxisFractionDigits } from "../../../../components/chart/y-axis-fraction-digits";
 import "../../../../components/ha-card";
 import type { EnergyData } from "../../../../data/energy";
 import {
   getEnergyDataCollection,
-  getPowerFromState,
   validateEnergyCollectionKey,
 } from "../../../../data/energy";
-import type { StatisticValue } from "../../../../data/recorder";
 import type { FrontendLocaleData } from "../../../../data/translation";
 import { SubscribeMixin } from "../../../../mixins/subscribe-mixin";
 import type { HomeAssistant } from "../../../../types";
 import type { LovelaceCard } from "../../types";
 import type { PowerSourcesGraphCardConfig } from "../types";
 import { hasConfigChanged } from "../../common/has-changed";
-import { getCommonOptions, fillLineGaps } from "./common/energy-chart-options";
+import { getCommonOptions } from "./common/energy-chart-options";
 import type { HaECOption } from "../../../../resources/echarts/echarts";
-import { hex2rgb } from "../../../../common/color/convert-color";
 import type { CustomLegendOption } from "../../../../components/chart/ha-chart-base";
+import { generatePowerSourcesGraphData } from "./power-sources-graph-data";
 
 @customElement("hui-power-sources-graph-card")
 export class HuiPowerSourcesGraphCard
@@ -170,254 +166,21 @@ export class HuiPowerSourcesGraphCard
   );
 
   private async _getStatistics(energyData: EnergyData): Promise<void> {
-    const datasets: LineSeriesOption[] = [];
-    this._legendData = [];
-
-    const statIds = {
-      solar: {
-        stats: [] as string[],
-        color: "--energy-solar-color",
-        name: this.hass.localize(
-          "ui.panel.lovelace.cards.energy.power_graph.solar"
-        ),
-      },
-      grid: {
-        stats: [] as string[],
-        color: "--energy-grid-consumption-color",
-        name: this.hass.localize(
-          "ui.panel.lovelace.cards.energy.power_graph.grid"
-        ),
-      },
-      battery: {
-        stats: [] as string[],
-        color: "--energy-battery-out-color",
-        name: this.hass.localize(
-          "ui.panel.lovelace.cards.energy.power_graph.battery"
-        ),
-      },
-    };
-
-    const computedStyles = getComputedStyle(this);
-
-    let yMin = Infinity;
-    let yMax = -Infinity;
-    const trackY = (v: number) => {
-      if (v < yMin) yMin = v;
-      if (v > yMax) yMax = v;
-    };
-
-    for (const source of energyData.prefs.energy_sources) {
-      if (source.type === "solar") {
-        if (source.stat_rate) {
-          statIds.solar.stats.push(source.stat_rate);
-        }
-        continue;
-      }
-
-      if (source.type === "battery") {
-        if (source.stat_rate) {
-          statIds.battery.stats.push(source.stat_rate);
-        }
-        continue;
-      }
-
-      if (source.type === "grid") {
-        if (source.stat_rate) {
-          statIds.grid.stats.push(source.stat_rate);
-        }
-      }
-    }
-    const commonSeriesOptions: LineSeriesOption = {
-      type: "line",
-      smooth: 0.4,
-      lineStyle: {
-        width: 1,
-      },
-    };
-
-    const now = Date.now();
-    const seriesData: Record<
-      string,
-      {
-        colorHex: string;
-        rgb: [number, number, number];
-        positive: [number, number][];
-        negative: [number, number][];
-      }
-    > = {};
-
-    Object.keys(statIds).forEach((key) => {
-      if (statIds[key].stats.length) {
-        const colorHex = computedStyles.getPropertyValue(statIds[key].color);
-        const rgb = hex2rgb(colorHex);
-        // Echarts is supposed to handle that but it is bugged when you use it together with stacking.
-        // The interpolation breaks the stacking, so this positive/negative is a workaround
-        const { positive, negative } = this._processData(
-          statIds[key].stats.map((id: string) => {
-            const stats = energyData.stats[id] ?? [];
-            if (isSameDay(now, this._start) && isSameDay(now, this._end)) {
-              // Append current state if we are showing today
-              const currentStateWatts = getPowerFromState(this.hass.states[id]);
-              if (currentStateWatts !== undefined) {
-                // getPowerFromState returns power in W; convert to kW for this graph
-                stats.push({
-                  start: now,
-                  end: now,
-                  mean: currentStateWatts / 1000,
-                });
-              }
-            }
-            return stats;
-          }),
-          trackY
-        );
-
-        seriesData[key] = { colorHex, rgb, positive, negative };
-      }
+    const result = generatePowerSourcesGraphData({
+      localize: this.hass.localize,
+      states: this.hass.states,
+      energyData,
+      computedStyles: getComputedStyle(this),
+      start: this._start,
+      end: this._end,
+      now: Date.now(),
     });
 
-    const pushSeries = (
-      key: string,
-      data: [number, number][],
-      stack: "positive" | "negative",
-      z: number
-    ) => {
-      const { colorHex, rgb } = seriesData[key];
-
-      datasets.push({
-        ...commonSeriesOptions,
-        id: stack === "positive" ? key : `${key}-negative`,
-        name: statIds[key].name,
-        color: colorHex,
-        stack,
-        areaStyle: {
-          color: new LinearGradient(
-            0,
-            stack === "positive" ? 0 : 1,
-            0,
-            stack === "positive" ? 1 : 0,
-            [
-              {
-                offset: 0,
-                color: `rgba(${rgb[0]}, ${rgb[1]}, ${rgb[2]}, 0.75)`,
-              },
-              {
-                offset: 1,
-                color: `rgba(${rgb[0]}, ${rgb[1]}, ${rgb[2]}, 0.25)`,
-              },
-            ]
-          ),
-        },
-        data,
-        z,
-      });
-    };
-
-    // Draw in reverse order so 0 value lines are overwritten
-    ["solar", "battery", "grid"].forEach((key, i) => {
-      if (seriesData[key]) {
-        pushSeries(key, seriesData[key].positive, "positive", 3 - i);
-      }
-    });
-
-    // Draw in reverse order but above positive series
-    ["battery", "grid"].forEach((key, i) => {
-      if (seriesData[key]) {
-        pushSeries(key, seriesData[key].negative, "negative", 4 - i);
-      }
-    });
-
-    Object.keys(statIds).forEach((key) => {
-      if (seriesData[key]) {
-        const { colorHex, rgb } = seriesData[key];
-
-        this._legendData!.push({
-          id: key,
-          secondaryIds: key !== "solar" ? [`${key}-negative`] : [],
-          name: statIds[key].name,
-          itemStyle: {
-            color: `rgba(${rgb[0]}, ${rgb[1]}, ${rgb[2]}, 0.75)`,
-            borderColor: colorHex,
-          },
-        });
-      }
-    });
-
-    this._start = energyData.start;
-    this._end = energyData.end || endOfToday();
-
-    this._chartData = fillLineGaps(datasets);
-    this._yAxisFractionDigits = computeYAxisFractionDigits(yMin, yMax);
-
-    const usageData: NonNullable<LineSeriesOption["data"]> = [];
-    this._chartData[0]?.data!.forEach((item, i) => {
-      // fillLineGaps ensures all datasets have the same x values
-      const x =
-        typeof item === "object" && "value" in item!
-          ? item.value![0]
-          : item![0];
-      let sum = 0;
-      this._chartData.forEach((dataset) => {
-        const y =
-          typeof dataset.data![i] === "object" && "value" in dataset.data![i]!
-            ? dataset.data![i].value![1]
-            : dataset.data![i]![1];
-        sum += y as number;
-      });
-      // Consumption can't be negative; sources unaccounted for in the
-      // configuration (e.g. solar exporting to grid without a configured
-      // solar source) would otherwise drag the usage line below zero.
-      usageData[i] = [x, Math.max(0, sum)];
-    });
-    this._chartData.push({
-      ...commonSeriesOptions,
-      id: "usage",
-      name: this.hass.localize(
-        "ui.panel.lovelace.cards.energy.power_graph.usage"
-      ),
-      color: computedStyles.getPropertyValue("--primary-text-color"),
-      lineStyle: {
-        type: [7, 2],
-        width: 1.5,
-      },
-      data: usageData,
-      z: 5,
-    });
-    this._legendData!.push({
-      id: "usage",
-      name: this.hass.localize(
-        "ui.panel.lovelace.cards.energy.power_graph.usage"
-      ),
-      itemStyle: {
-        color: computedStyles.getPropertyValue("--primary-text-color"),
-      },
-    });
-  }
-
-  private _processData(stats: StatisticValue[][], trackY: (v: number) => void) {
-    const data: Record<number, number[]> = {};
-    stats.forEach((statSet) => {
-      statSet.forEach((point) => {
-        if (point.mean == null) {
-          return;
-        }
-        const x = (point.start + point.end) / 2;
-        data[x] = [...(data[x] ?? []), point.mean];
-      });
-    });
-    const positive: [number, number][] = [];
-    const negative: [number, number][] = [];
-    Object.entries(data).forEach(([x, y]) => {
-      const ts = Number(x);
-      const sumY = y.reduce((a, b) => a + b, 0);
-      const pos = Math.max(0, sumY);
-      const neg = Math.min(0, sumY);
-      positive.push([ts, pos]);
-      negative.push([ts, neg]);
-      trackY(pos);
-      trackY(neg);
-    });
-    return { positive, negative };
+    this._legendData = result.legendData;
+    this._start = result.start;
+    this._end = result.end;
+    this._chartData = result.chartData;
+    this._yAxisFractionDigits = result.yAxisFractionDigits;
   }
 
   static styles = css`
